requirements.txt
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cs50==9.0.0
finnhub-python==2.4.13
Flask==2.1.1
Flask-Session==0.4.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
multitasking==0.0.10
numeral==0.1.0.17
numerize==0.12
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.1
yfinance==0.1.70
